refactor(database): modularizar queries e operações em submódulos focados no pacote src.core.db
