+++ src/core/database.py
@@ -1,322 +1,9 @@
-import os
-import traceback
-import uuid
-
-import streamlit as st
-
-from src.config import (
-    LIMITE_TEMAS,
-    MAX_CHUNCK,
-    SEMANTICA_THRESHOLD,
-    TAMANHO_VETOR_SEMANTICO,
-    get_secret,
-    logger,
+from src.core.db.client import get_db_client
+from src.core.db.sessions import salvar_sessao, excluir_dados_sessao
+from src.core.db.logs import salvar_log_chat, salvar_erro
+from src.core.db.reports import salvar_report, get_categorias_erro
+from src.core.db.retrieval import (
+    buscar_referencias_db,
+    buscar_chunks_por_topico,
+    recuperar_contexto_inteligente,
 )
-from supabase import Client, create_client
-
-
-@st.cache_resource
-def get_db_client() -> Client:
-    try:
-        url = get_secret("supabase.url")
-        key = get_secret("supabase.key")
-
-        if not url or not key:
-            logger.error("Credenciais do Supabase não encontradas.")
-            return None
-
-        return create_client(url, key)
-    except Exception as e:
-        st.error(f"Erro ao conectar no banco de dados: {e}")
-        return None
-
-
-def salvar_sessao(session_id):
-    client = get_db_client()
-    if not client:
-        logger.error("Não foi possível conectar com o banco de dados.")
-        return
-    try:
-        registro_sessao = {"session_id": session_id}
-        client.table("sessions").insert(registro_sessao).execute()
-    except Exception as e:
-        logger.error(f"⚠️ Erro ao tentar registrar sessão no banco de dados: {e}")
-
-
-def salvar_log_chat(
-    session_id, git_version, prompt, response, fonte_info, lista_kb_ids=None
-):
-
-    client = get_db_client()
-    if not client:
-        logger.error("❌ Não foi possível conectar com o banco de dados.")
-        return
-
-    try:
-        data_log = {
-            "session_id": session_id,
-            "prompt": prompt,
-            "response": str(response),
-            "git_version": git_version,
-        }
-
-        res = client.table("chat_logs").insert(data_log).execute()
-
-        if not res.data:
-            logger.warning("⚠️ Erro: Log salvo mas sem retorno de ID.")
-            return
-
-        novo_log_id = res.data[0]["chat_id"]
-
-        if lista_kb_ids and len(lista_kb_ids) > 0:
-            dados_relacao = []
-            for item in lista_kb_ids:
-                if isinstance(item, dict):
-                    kb_id = item.get("kb_id")
-                    similarity = item.get("similarity")
-                else:
-                    kb_id = item
-                    similarity = None
-
-                if kb_id:
-                    row_data = {
-                        "chat_id": novo_log_id,
-                        "kb_id": str(kb_id),
-                    }
-                    if similarity is not None:
-                        row_data["similarity"] = similarity
-
-                    dados_relacao.append(row_data)
-
-            if dados_relacao:
-                try:
-                    res_kb = (
-                        client.table("chat_logs_kb").insert(dados_relacao).execute()
-                    )
-                except Exception as e_kb:
-                    logger.error(f"❌ ERRO ao inserir em chat_logs_kb: {e_kb}")
-                    logger.error(f"❌ Dados tentados: {dados_relacao}")
-            else:
-                logger.info("⚠️ Nenhuma relação válida para inserir.")
-        else:
-            pass
-
-    except Exception as e:
-        logger.error(f"❌ ERRO ao salvar log: {type(e).__name__}")
-        logger.error(f"❌ Mensagem de erro: {e}", exc_info=True)
-
-
-def salvar_erro(session_id, git_version, error_msg):
-    # 1. Log imediato no sistema (Console/Arquivo) com Traceback
-    logger.error(f"🔥 Exceção capturada e registrada: {error_msg}", exc_info=True)
-
-    client = get_db_client()
-    if not client:
-        logger.error("Falha ao obter cliente DB para salvar log de erro.")
-        return "ERRO-DB"
-    try:
-        error_id = str(uuid.uuid4())[:8]
-
-        data = {
-            "error_id": error_id,
-            "error_message": str(error_msg),
-            "session_id": session_id,
-            "git_version": git_version,
-        }
-        client.table("error_logs").insert(data).execute()
-        return error_id
-    except Exception as e:
-        logger.error(f"⚠️ CRÍTICO: Falha ao salvar o erro no Supabase: {e}")
-        return "N/A"
-
-
-def salvar_report(session_id, git_version, history_text, category_id, comment):
-    client = get_db_client()
-
-    try:
-        if not client:
-            return False
-        data = {
-            "session_id": session_id,
-            "git_version": git_version,
-            "chat_history": history_text,
-            "category_id": category_id,
-            "comment": comment,
-        }
-        client.table("user_reports").insert(data).execute()
-        return True
-    except Exception as e:
-        logger.error(f"⚠️ Erro ao salvar report: {e}")
-        return False
-
-
-def get_categorias_erro():
-    client = get_db_client()
-    try:
-        if not client:
-            return False
-        response = client.table("report_categories").select("id", "label").execute()
-        return response.data if response.data else []
-
-    except Exception as e:
-        logger.error(f"⚠️ Erro ao buscar categorias: {e}")
-        return []
-
-
-def buscar_referencias_db(
-    vector_embedding,
-    threshold=SEMANTICA_THRESHOLD,
-    limit=LIMITE_TEMAS,
-    filter_topic=None,
-):
-    client = get_db_client()
-    if not client:
-        logger.error("⚠️ Erro: Cliente Supabase não inicializado.")
-        return []
-    try:
-        if len(vector_embedding) != TAMANHO_VETOR_SEMANTICO:
-            logger.error(
-                f"⚠️ Erro de Dimensão: O vetor gerado tem {len(vector_embedding)} dimensões, mas o banco espera {TAMANHO_VETOR_SEMANTICO}."
-            )
-            return []
-
-        params = {
-            "query_embedding": vector_embedding,
-            "match_threshold": threshold,
-            "match_count": limit,
-            "filter_topic": filter_topic,
-        }
-
-        response = client.rpc("match_knowledge_base", params).execute()
-
-        if response.data:
-            return response.data
-        else:
-            logger.info("⚠️ Nenhum match encontrado na base com esse threshold.")
-        return []
-
-    except Exception as e:
-        logger.critical(f"❌ Erro CRÍTICO na busca vetorial (Supabase): {e}")
-        return []
-
-
-def buscar_chunks_por_topico(topico_alvo, limit=30):
-    client = get_db_client()
-    if not client:
-        return []
-    try:
-        response = (
-            client.table("knowledge_base")
-            .select("kb_id, descricao")
-            .eq("topico", topico_alvo)
-            .limit(limit)
-            .execute()
-        )
-        return response.data if response.data else []
-    except Exception as e:
-        logger.error(f"❌ Erro ao buscar tópico completo: {e}")
-        return []
-
-
-def recuperar_contexto_inteligente(vector_embedding):
-    client = get_db_client()
-    if not client:
-        logger.error("⚠️ Erro: Cliente Supabase não inicializado.")
-        return [], "Erro DB", []
-    resultados_iniciais = buscar_referencias_db(
-        vector_embedding, SEMANTICA_THRESHOLD, LIMITE_TEMAS, None
-    )
-    if not resultados_iniciais:
-        return None, "Nenhuma referencia encontrada na base de conhecimento.", None
-
-    def _gerar_fallback_top5():
-        top_5 = resultados_iniciais[:5]
-        contexto = [item["descricao"] for item in top_5]
-        ids_usados = []
-        for item in top_5:
-            kid = item.get("kb_id") or item.get("id")
-            if kid:
-                ids_usados.append(
-                    {"kb_id": kid, "similarity": item.get("similarity")}
-                )
-        return contexto, ids_usados
-
-    contagem_topicos = {}
-
-    for item in resultados_iniciais:
-        topico = item.get("topico")
-        if topico:
-            contagem_topicos[topico] = contagem_topicos.get(topico, 0) + 1
-
-    contexto_final = []
-    lista_ids_usados = []
-    fonte_origem = "Busca por similaridade (Fragmentos)"
-
-    if not contagem_topicos:
-        contexto_final, lista_ids_usados = _gerar_fallback_top5()
-        return "\n---\n".join(contexto_final), fonte_origem, lista_ids_usados
-
-    topico_vencedor = max(contagem_topicos, key=contagem_topicos.get)
-    votos = contagem_topicos[topico_vencedor]
-
-    if votos >= 3:
-        logger.info(
-            f"🚀 Estratégia: Contexto Expandido para o tópico '{topico_vencedor}'"
-        )
-        try:
-            dados = buscar_chunks_por_topico(topico_vencedor, limit=MAX_CHUNCK)
-
-            contexto_final = [row["descricao"] for row in dados]
-            for row in dados:
-                kid = row.get("kb_id") or row.get("id")
-                if kid:
-                    lista_ids_usados.append({"kb_id": kid, "similarity": None})
-
-            fonte_origem = f"Contexto Completo: {topico_vencedor}"
-
-        except Exception as e:
-            logger.warning(f"⚠️ Erro ao expandir contexto: {e}. Usando fallback.")
-            contexto_final, lista_ids_usados = _gerar_fallback_top5()
-
-    else:
-        logger.info(f"🔍 Estratégia: Tópicos mistos (Vencedor '{topico_vencedor}')")
-        fonte_origem = f"Tópicos mistos (Vencedor: {topico_vencedor})"
-        contexto_final, lista_ids_usados = _gerar_fallback_top5()
-
-    return "\n---\n".join(contexto_final), fonte_origem, lista_ids_usados
-
-
-def excluir_dados_sessao(session_id: str) -> bool:
-    """
-    Exclui permanentemente todos os registros vinculados ao session_id 
-    nas tabelas chat_logs_kb, chat_logs, user_reports, error_logs e sessions 
-    para cumprir o Art. 18 da LGPD.
-    """
-    client = get_db_client()
-    if not client:
-        logger.error("Não foi possível conectar ao banco de dados para excluir dados.")
-        return False
-    try:
-        # 1. Obtém os chat_ids dessa sessão para excluir as referências em chat_logs_kb
-        res_logs = client.table("chat_logs").select("chat_id").eq("session_id", session_id).execute()
-        if res_logs.data:
-            chat_ids = [row["chat_id"] for row in res_logs.data]
-            client.table("chat_logs_kb").delete().in_("chat_id", chat_ids).execute()
-            
-        # 2. Deleta os logs de chat
-        client.table("chat_logs").delete().eq("session_id", session_id).execute()
-        
-        # 3. Deleta os relatórios de usuário
-        client.table("user_reports").delete().eq("session_id", session_id).execute()
-        
-        # 4. Deleta os logs de erro
-        client.table("error_logs").delete().eq("session_id", session_id).execute()
-        
-        # 5. Deleta a sessão em si
-        client.table("sessions").delete().eq("session_id", session_id).execute()
-        
-        logger.info(f"Dados da sessão {session_id} foram excluídos permanentemente (LGPD Art. 18).")
-        return True
-    except Exception as e:
-        logger.error(f"Erro ao excluir dados da sessão {session_id} no banco de dados: {e}")
-        return False
