+++ src/core/db/__init__.py
@@ -1,0 +1,1 @@
+# Pacote contendo submódulos de banco de dados do Supabase
